Bugs fixed
#20 - Background Color Change
When the program checks which key has been newly entered, it uses the same register used to delete the screen.
We then reset this register to zero at the start of our loop, so that the screen is always black.

#21 - Key reading problem
Direct Pac-Man launch to prevent different versions of Pac-Man from launching directly.